chore: Refactor mappers to use spread operator for data and entity

+++ src/config/reservation/reservation.service.impl.ts
@@ -1,21 +1,28 @@
-import { Injectable } from '@nestjs/common'
-import { ExternalReservationCreateDto } from '@/module/reservation/dto/reservation-create.dto'
+import { ExternalReservationCreateDto, ReservationCreateDto } from '@/module/reservation/dto/reservation-create.dto'
+import { Injectable, NotFoundException, Param } from '@nestjs/common'
 
 import { Filter, Pageable, PaginationRequest } from '@/data/util'
-import { ReservationService } from '@/module/reservation/service/reservation.service'
-import { Reservation } from '@/module/reservation/model/reservation.model'
 import { ReservationStatus } from '@/module/reservation/model/reservation-status.enum'
+import { Reservation } from '@/module/reservation/model/reservation.model'
 import { ReservationRepository } from '@/module/reservation/repository/reservation.repository'
+import { ReservationService } from '@/module/reservation/service/reservation.service'
+import { RestaurantRepository } from '@/data/repository/restaurant.repository'
+import { UserRepository } from '@/data/repository/user.repository'
+import { ParameterException } from '@/exception/parameter-exception'
 
 
 @Injectable()
 export class ReservationServiceImpl implements ReservationService {
     constructor(
-        private readonly reservationRepository: ReservationRepository
+        private readonly reservationRepository: ReservationRepository,
+        private readonly restaurantRepository: RestaurantRepository,
+        private readonly userRepository: UserRepository,
     ) { }
 
     async getReservationById(reservationId: string): Promise<Reservation> {
-        return this.reservationRepository.findOne(reservationId)
+        const reservation = await this.reservationRepository.findOne(reservationId)
+        if (!reservation) return null;
+        return reservation;
     }
 
     getReservationByUserId(
@@ -44,16 +51,51 @@ export class ReservationServiceImpl implements ReservationService {
         })
     }
 
-    async createReservation(
-        reservation: ExternalReservationCreateDto
+
+    async #createReservation(
+        reservation: Omit<Reservation, "id" | "createdAt" | "updatedAt">
     ): Promise<Reservation> {
-        return this.reservationRepository.createOne(reservation.toEntity())
+
+        // if null/undefined values
+        if (!reservation.date) {
+            throw new ParameterException("reservation.date", 'Date is required')
+        }
+        if (!reservation.nb_people) {
+            throw new ParameterException("reservation.nb_people", 'Number of people is required')
+        }
+        if (!reservation.restaurantId) {
+            throw new ParameterException("reservation.restaurantId", 'Restaurant id is required')
+        }
+        if (!reservation.status) {
+            reservation.status = ReservationStatus.PENDING
+        }
+
+        // check that the restaurant exists
+        if (!(await this.restaurantRepository.findOne(reservation.restaurantId))) {
+            throw new ParameterException("reservation.restaurantId", 'Restaurant not found')
+        }
+
+        // check that the user exists
+        if (
+            reservation.userId &&
+            !(await this.userRepository.findOne(reservation.userId))) {
+            throw new ParameterException("reservation.userId", 'User not found')
+        }
+
+        return this.reservationRepository.createOne(reservation)
     }
+
+    async createReservation(
+        reservation: ReservationCreateDto
+    ): Promise<Reservation> {
+        return this.#createReservation(reservation.toEntity())
+    }
+
 
     async createExternalReservation(
         reservation: ExternalReservationCreateDto
     ): Promise<Reservation> {
-        return this.reservationRepository.createOne(reservation.toEntity())
+        return this.#createReservation(reservation.toEntity())
     }
 
     async updateReservationStatus(
@@ -62,9 +104,13 @@ export class ReservationServiceImpl implements ReservationService {
     ): Promise<Reservation> {
         const reservation =
             await this.reservationRepository.findOne(reservationId)
-        if (!reservation) throw new Error('Reservation not found')
-        if (reservation.status !== ReservationStatus.PENDING)
-            throw new Error('Reservation status cannot be updated')
+
+        if (!reservation) {
+            throw new NotFoundException('Reservation not found')
+        }
+        if (reservation.status !== ReservationStatus.PENDING) {
+            throw new ParameterException('reservationId', 'Reservation can only be updated if it is pending')
+        }
 
         reservation.status = status
         return this.reservationRepository.updateOne(reservationId, reservation)
@@ -73,17 +119,37 @@ export class ReservationServiceImpl implements ReservationService {
     async updateReservation(
         reservation: Partial<Reservation>
     ): Promise<Reservation> {
-        const originalReservation = await this.reservationRepository.findOne(
+        const $reservation = await this.reservationRepository.findOne(
             reservation.id
         )
-        if (!originalReservation) throw new Error('Reservation not found')
-        if (originalReservation.status !== ReservationStatus.PENDING)
-            throw new Error('Reservation status cannot be updated')
+        // check reservation exists
+        if (!$reservation) {
+            throw new NotFoundException('Reservation not found')
+        }
+
+        // a reservation can only be updated if it is pending
+        if ($reservation.status !== ReservationStatus.PENDING) {
+            throw new ParameterException('reservationId', `Reservation can only be updated if it is pending${$reservation.status}`)
+        }
+
+        // check user exists
+        if (reservation.userId && !(await this.userRepository.findOne(reservation.userId))) {
+            throw new ParameterException("reservation.userId", 'User not found')
+        }
+
+        // check restaurant exists
+        if (reservation.restaurantId && !(await this.restaurantRepository.findOne(reservation.restaurantId))) {
+            throw new ParameterException("reservation.restaurantId", 'Restaurant not found')
+        }
 
         return this.reservationRepository.updateOne(reservation.id, reservation)
     }
 
     async deleteReservation(reservationId: string): Promise<void> {
+        if (!await this.reservationRepository.findOne(reservationId)) {
+            throw new NotFoundException('Reservation not found')
+        }
+
         return this.reservationRepository.deleteOne(reservationId)
     }
 }
